+++ README.md
@@ -1,5 +1,17 @@
-# サーバーの起動方法
-`$ go run src/app/server.go`
+# CAMPHOR- COIN
+
+## インストール
+
+```
+$ go get github.com/CAMPHOR-COIN/camphor-coin
+```
+
+## サーバーの起動方法
+
+```
+$ cd $GOROOT/src/github.com/CAMPHOR-COIN/camphor-coin
+$ go run server.go
+```
 
 上記コマンドで8080番でサーバーが起動。 `/blockchain`を叩くと最新のBlockchainが確認できる
