+++ interfaces/controllers/blockchain_controller.go
@@ -1,0 +1,24 @@
+package controllers
+
+import (
+	"github.com/CAMPHOR-COIN/camphor-coin/domain"
+)
+
+type BlockchainController struct {}
+
+func NewBlockchainController() *BlockchainController {
+	return &BlockchainController{}
+}
+
+func (controller *BlockchainController) Show(c Context) {
+	// TODO DBから取得した最新のチェーンを利用する - 20180204 andoshin11
+	blockchain := domain.GetBlockchain()
+
+	// blockchain, err := controller.interactor.GetBlockchain()
+	// if err != nil {
+	//   c.JSON(500, NewError(err))
+	//   return
+	// }
+
+	c.JSON(200, blockchain)
+}
