--- src/app/interfaces/controllers/context.go
@@ -1,8 +1,0 @@
-package controllers
-
-type Context interface {
-  Param(string) string
-  Bind(interface{}) error
-  Status(int)
-  JSON(int, interface{})
-}
