--- src/app/server.go
@@ -1,8 +1,0 @@
-package main
-
-import "app/infrastructure"
-
-func main() {
-	// サーバーの起動
-	infrastructure.Router.Run()
-}
